+++ fltk/examples/flex.rs
@@ -16,16 +16,16 @@ fn main() {
 
 fn buttons_panel(parent: &mut group::Flex) {
     frame::Frame::default();
-    let mut w = frame::Frame::default().with_label("Welcome to Flex Login");
+    let w = frame::Frame::default().with_label("Welcome to Flex Login");
 
     let mut urow = group::Flex::default().row();
     {
         frame::Frame::default()
             .with_label("Username:")
             .with_align(enums::Align::Inside | enums::Align::Right);
-        let mut username = input::Input::default();
+        let username = input::Input::default();
 
-        urow.set_size(&mut username, 180);
+        urow.set_size(&username, 180);
         urow.end();
     }
 
@@ -34,35 +34,35 @@ fn buttons_panel(parent: &mut group::Flex) {
         frame::Frame::default()
             .with_label("Password:")
             .with_align(enums::Align::Inside | enums::Align::Right);
-        let mut password = input::Input::default();
+        let password = input::Input::default();
 
-        prow.set_size(&mut password, 180);
+        prow.set_size(&password, 180);
         prow.end();
     }
 
-    let mut pad = frame::Frame::default();
+    let pad = frame::Frame::default();
 
     let mut brow = group::Flex::default().row();
     {
         frame::Frame::default();
-        let mut reg = create_button("Register");
-        let mut login = create_button("Login");
+        let reg = create_button("Register");
+        let login = create_button("Login");
 
-        brow.set_size(&mut reg, 80);
-        brow.set_size(&mut login, 80);
+        brow.set_size(&reg, 80);
+        brow.set_size(&login, 80);
         brow.end();
     }
 
-    let mut b = frame::Frame::default();
+    let b = frame::Frame::default();
 
     frame::Frame::default();
 
-    parent.set_size(&mut w, 60);
-    parent.set_size(&mut urow, 30);
-    parent.set_size(&mut prow, 30);
-    parent.set_size(&mut pad, 1);
-    parent.set_size(&mut brow, 30);
-    parent.set_size(&mut b, 30);
+    parent.set_size(&w, 60);
+    parent.set_size(&urow, 30);
+    parent.set_size(&prow, 30);
+    parent.set_size(&pad, 1);
+    parent.set_size(&brow, 30);
+    parent.set_size(&b, 30);
 }
 
 fn middle_panel(parent: &mut group::Flex) {
@@ -71,7 +71,7 @@ fn middle_panel(parent: &mut group::Flex) {
     let mut frame = frame::Frame::default().with_label("Image");
     frame.set_frame(enums::FrameType::BorderBox);
     frame.set_color(enums::Color::from_rgb(0, 200, 0));
-    let mut spacer = frame::Frame::default();
+    let spacer = frame::Frame::default();
 
     let mut bp = group::Flex::default().column();
     buttons_panel(&mut bp);
@@ -79,9 +79,9 @@ fn middle_panel(parent: &mut group::Flex) {
 
     frame::Frame::default();
 
-    parent.set_size(&mut frame, 200);
-    parent.set_size(&mut spacer, 10);
-    parent.set_size(&mut bp, 300);
+    parent.set_size(&frame, 200);
+    parent.set_size(&spacer, 10);
+    parent.set_size(&bp, 300);
 }
 
 fn main_panel(parent: &mut group::Flex) {
@@ -93,7 +93,7 @@ fn main_panel(parent: &mut group::Flex) {
 
     frame::Frame::default();
 
-    parent.set_size(&mut mp, 200);
+    parent.set_size(&mp, 200);
 }
 
 fn create_button(caption: &str) -> button::Button {